fix up some fn sigs

+++ include/onex-kernel/log.h
@@ -28,9 +28,11 @@ extern bool log_to_rtt;
 
 void log_init(properties* config);
 bool log_loop();
+
 #define log_write(...)     log_write_current_file_line(0,0, __VA_ARGS__)
 #define log_write_src(...) log_write_current_file_line(__FILE__, __LINE__, __VA_ARGS__)
-int  log_write_current_file_line(char* file, uint32_t line, const char* fmt, ...);
+int16_t log_write_current_file_line(char* file, uint32_t line, const char* fmt, ...);
+
 void log_flash(uint8_t r, uint8_t g, uint8_t b);
 void log_flush();
